fix: code structure
